refactoring DQN SYNC + laneDetector with regression in step() method
