+++ rl_studio/agents/auto_carla/train_followlane_dqn_carla_tf.py
@@ -33,7 +33,14 @@ import torch
 from tqdm import tqdm
 import sys
 
-# from rl_studio.agents.auto_carla.sources import settings
+from rl_studio.agents.auto_carla.actors_sensors import (
+    NewCar,
+    CameraRGBSensor,
+    CameraRedMaskSemanticSensor,
+    # LaneDetector,
+)
+from rl_studio.agents.auto_carla.carla_env import CarlaEnv
+
 from rl_studio.agents.f1.loaders import (
     LoadAlgorithmParams,
     LoadEnvParams,
@@ -155,783 +162,6 @@ print(len(gpus), "Physical GPUs,", len(logical_gpus), "Logical GPUs")
 
 # Init Ray
 # ray.init(ignore_reinit_error=True)
-
-
-##################################################################################
-#
-# vars
-##################################################################################
-
-correct_normalized_distance = {
-    20: -0.07,
-    30: -0.1,
-    40: -0.13,
-    50: -0.17,
-    60: -0.2,
-    70: -0.23,
-    80: -0.26,
-    90: -0.3,
-    100: -0.33,
-    110: -0.36,
-    120: -0.4,
-    130: -0.42,
-    140: -0.46,
-    150: -0.49,
-    160: -0.52,
-    170: -0.56,
-    180: -0.59,
-    190: -0.62,
-    200: -0.65,
-    210: -0.69,
-    220: -0.72,
-}
-correct_pixel_distance = {
-    20: 343,
-    30: 353,
-    40: 363,
-    50: 374,
-    60: 384,
-    70: 394,
-    80: 404,
-    90: 415,
-    100: 425,
-    110: 436,
-    120: 446,
-    130: 456,
-    140: 467,
-    150: 477,
-    160: 488,
-    170: 498,
-    180: 508,
-    190: 518,
-    200: 528,
-    210: 540,
-    220: 550,
-}
-
-
-# ==============================================================================
-# -- class CarlaEnv -------------------------------------------------------------
-# ==============================================================================
-
-
-class CarlaEnv(gym.Env):
-    def __init__(self, car, sensor_camera_rgb, sensor_camera_lanedetector, config):
-        super(CarlaEnv, self).__init__()
-        ## --------------- init env
-        # FollowLaneEnv.__init__(self, **config)
-        ## --------------- init class variables
-        FollowLaneCarlaConfig.__init__(self, **config)
-
-        # print(f"{config =}")
-
-        self.image_raw_from_topic = None
-        self.image_camera = None
-        # self.sensor = config["sensor"]
-
-        # Image
-        self.image_resizing = config["image_resizing"] / 100
-        self.new_image_size = config["new_image_size"]
-        self.raw_image = config["raw_image"]
-        self.height = int(config["height_image"] * self.image_resizing)
-        self.width = int(config["width_image"] * self.image_resizing)
-        self.center_image = int(config["center_image"] * self.image_resizing)
-        self.num_regions = config["num_regions"]
-        self.pixel_region = int(self.center_image / self.num_regions) * 2
-        # self.telemetry_mask = config["telemetry_mask"]
-        self.poi = config["x_row"][0]
-        self.image_center = None
-        self.right_lane_center_image = config["center_image"] + (
-            config["center_image"] // 2
-        )
-        self.lower_limit = config["lower_limit"]
-
-        # States
-        self.state_space = config["states"]
-        self.states_entry = config["states_entry"]
-        if self.state_space == "spn":
-            self.x_row = [i for i in range(1, int(self.height / 2) - 1)]
-        else:
-            self.x_row = config["x_row"]
-
-        # Actions
-        self.actions_space = config["action_space"]
-        self.actions = config["actions"]
-
-        # Rewards
-        self.reward_function = config["reward_function"]
-        self.rewards = config["rewards"]
-        self.min_reward = config["min_reward"]
-
-        # Pose
-        self.town = config["town"]
-        self.random_pose = config["random_pose"]
-        self.alternate_pose = config["alternate_pose"]
-        self.init_pose_number = config["init_pose_number"]
-        self.finish_pose_number = config["finish_pose_number"]
-        self.start_alternate_pose = config["start_alternate_pose"]
-        self.finish_alternate_pose = config["finish_alternate_pose"]
-
-        # print(f"{self.alternate_pose =}\n")
-        # print(f"{self.start_alternate_pose =}\n")
-
-        self.waypoints_meters = config["waypoints_meters"]
-        self.waypoints_init = config["waypoints_init"]
-        self.waypoints_target = config["waypoints_target"]
-        self.waypoints_lane_id = config["waypoints_lane_id"]
-        self.waypoints_road_id = config["waypoints_road_id"]
-        self.max_target_waypoint_distance = config["max_target_waypoint_distance"]
-
-        ###############################################################
-        #
-        # gym/stable-baselines3 interface
-        ###############################################################
-
-        ######## Actions Gym based
-        print(f"{self.state_space = } and {self.actions_space =}")
-        # print(f"{len(self.actions) =}")
-        # print(f"{type(self.actions) =}")
-        # print(f"{self.actions =}")
-        # Discrete Actions
-        if self.actions_space == "carla_discrete":
-            self.action_space = spaces.Discrete(len(self.actions))
-        else:  # Continuous Actions
-            actions_to_array = np.array(
-                [list(self.actions["v"]), list(self.actions["w"])]
-            )
-            print(f"{actions_to_array =}")
-            print(f"{actions_to_array[0] =}")
-            print(f"{actions_to_array[1] =}")
-            self.action_space = spaces.Box(
-                low=actions_to_array[0],
-                high=actions_to_array[1],
-                shape=(2,),
-                dtype=np.float32,
-            )
-            print(f"{self.action_space.low = }")
-            print(f"{self.action_space.high = }")
-            print(f"{self.action_space.low[0] = }")
-            print(f"{self.action_space.high[0] = }")
-
-        print(f"{self.action_space =}")
-
-        ######## observations Gym based
-        # image
-        if self.state_space == "image":
-            self.observation_space = spaces.Box(
-                low=0, high=255, shape=(self.height, self.width, 3), dtype=np.uint8
-            )
-        else:  # Discrete observations vector = [0.0, 3.6, 8.9]
-            # TODO: change x_row for other list
-            self.observation_space = spaces.Discrete(len(self.x_row))  # temporary
-
-        print(f"{self.observation_space = }")
-
-        #########################################################################
-
-        # self.params = config
-        # print(f"{self.params =}\n")
-
-        self.car = car
-        self.sensor_camera_rgb = sensor_camera_rgb
-        self.sensor_camera_lanedetector = sensor_camera_lanedetector
-        self.params = {}
-        self.is_finish = None
-        self.dist_to_finish = None
-        self.collision_hist = []
-
-        self.right_line_in_pixels = None
-        self.ground_truth_pixel_values = None
-        self.dist_normalized = None
-        self.states = None
-        self.state_right_lines = None
-        self.drawing_lines_states = []
-        self.drawing_numbers_states = []
-
-        self.lane_changing_hist = []
-
-    #####################################################################################
-    #
-    #                                       RESET
-    #
-    #####################################################################################
-    def reset(self, seed=None, options=None):
-        """
-        state = vector / simplified perception
-        actions = discrete
-        """
-
-        ############################################################################
-        ########### --- calculating STATES
-
-        ##### EATA PARTE NOS SIRVE PARA ENCONTRAR EL CENTRO SOLO CON LA LINEA DERECHA
-        while self.sensor_camera_rgb.front_rgb_camera is None:
-            print(f"RESET() ----> {self.sensor_camera_rgb.front_rgb_camera = }")
-            time.sleep(0.2)
-
-        # mask = self.preprocess_image(self.front_red_mask_camera)
-        # mask = self.preprocess_image(self.sensor_camera.front_red_mask_camera)
-        # self.right_line_in_pixels = self.calculate_right_line(mask, self.x_row)
-
-        """
-        ##### EATA PARTE NOS SIRVE PARA ENCONTRAR EL CENTRO SOLO CON EL LANE DETECTOR
-        (
-            image_rgb_lanedetector,
-            left_mask,
-            right_mask,
-        ) = self.sensor_camera_lanedetector.detect(
-            self.sensor_camera_rgb.front_rgb_camera
-        )
-        mask = self.preprocess_image_lane_detector(image_rgb_lanedetector)
-        lane_centers_in_pixels, _, _ = self.calculate_lane_centers_with_lane_detector(
-            mask
-        )
-
-        pixels_in_state = mask.shape[1] / self.num_regions
-        self.states = [
-            int(value / pixels_in_state)
-            for _, value in enumerate(lane_centers_in_pixels)
-        ]
-        
-        """
-        self.states = [self.num_regions // 2 for i, _ in enumerate(self.x_row)]
-        print(f"\n\t{self.states =}")
-        # states_size = len(self.states)
-
-        """
-        AutoCarlaUtils.show_image(
-            "mask",
-            mask,
-            50,
-            500,
-        )
-
-        AutoCarlaUtils.show_image(
-            "LaneDetector RGB",
-            image_rgb_lanedetector[(image_rgb_lanedetector.shape[0] // 2) :],
-            600,
-            500,
-        )
-
-        AutoCarlaUtils.show_image(
-            "front RGB",
-            self.sensor_camera_rgb.front_rgb_camera[
-                (self.sensor_camera_rgb.front_rgb_camera.shape[0] // 2) :
-            ],
-            1250,
-            500,
-        )
-        """
-
-        # print(f"{states =} and {states_size =}")
-        return self.states  # , states_size
-
-    #####################################################################################
-    #
-    #                                       STEP
-    #
-    #####################################################################################
-    def step(self, action):
-        """
-        state: sp
-        actions: continuous
-        only right line
-        """
-
-        self.control_discrete_actions(action)
-        print(f"\n\tSTEP() POR AQUI")
-
-        ########### --- calculating STATES
-        mask = self.preprocess_image(self.sensor_camera_red_mask.front_red_mask_camera)
-
-        ########### --- Calculating center ONLY with right line
-        self.right_line_in_pixels = self.calculate_right_line(mask, self.x_row)
-
-        image_center = mask.shape[1] // 2
-        # dist = [
-        #    image_center - right_line_in_pixels[i]
-        #    for i, _ in enumerate(right_line_in_pixels)
-        # ]
-        self.dist_normalized = [
-            float((image_center - self.right_line_in_pixels[i]) / image_center)
-            for i, _ in enumerate(self.right_line_in_pixels)
-        ]
-
-        ## STATES
-
-        # pixels_in_state = mask.shape[1] // self.num_regions
-        # self.state_right_lines = [
-        #    i for i in range(1, mask.shape[1]) if i % pixels_in_state == 0
-        # ]
-        # self.states = [
-        #    int(value / pixels_in_state)
-        #    for _, value in enumerate(self.right_line_in_pixels)
-        # ]
-
-        # non regular states: 1 to n-2 in center. n-1 right, n left
-        size_lateral_states = 140
-        size_center_states = mask.shape[1] - (size_lateral_states * 2)
-        pixel_center_states = int(size_center_states / (self.num_regions - 2))
-
-        self.states = [
-            int(((value - size_lateral_states) / pixel_center_states) + 1)
-            if (mask.shape[1] - size_lateral_states) > value > size_lateral_states
-            else self.num_regions - 1
-            if value >= (mask.shape[1] - size_lateral_states)
-            else self.num_regions
-            for _, value in enumerate(self.right_line_in_pixels)
-        ]
-
-        # drawing lines and numbers states in image
-        self.drawing_lines_states = [
-            size_lateral_states + (i * pixel_center_states)
-            for i in range(1, self.num_regions - 1)
-        ]
-        # self.drawing_lines_states.append(size_lateral_states)
-        self.drawing_lines_states.insert(0, size_lateral_states)
-
-        self.drawing_numbers_states = [
-            i if i > 0 else self.num_regions for i in range(0, self.num_regions)
-        ]
-        # print(f"\n{self.drawing_lines_states}")
-        # print(f"\n{self.drawing_numbers_states}")
-
-        ## -------- Ending Step()...
-        done = False
-        ground_truth_normal_values = [
-            correct_normalized_distance[value] for i, value in enumerate(self.x_row)
-        ]
-        # reward, done = self.autocarlarewards.rewards_right_line(
-        #    dist_normalized, ground_truth_normal_values, self.params
-        # )
-        reward, done = self.autocarlarewards.rewards_sigmoid_only_right_line(
-            self.dist_normalized, ground_truth_normal_values
-        )
-
-        ## -------- ... or Finish by...
-        if len(self.collision_hist) > 0:  # crashed you, baby
-            done = True
-            # reward = -100
-            print(f"crash")
-
-        self.is_finish, self.dist_to_finish = AutoCarlaUtils.finish_fix_number_target(
-            self.params["location"],
-            self.finish_alternate_pose,
-            self.finish_pose_number,
-            self.max_target_waypoint_distance,
-        )
-        if self.is_finish:
-            print(f"Finish!!!!")
-            done = True
-
-        self.ground_truth_pixel_values = [
-            correct_pixel_distance[value] for i, value in enumerate(self.x_row)
-        ]
-
-        return self.states, reward, done, {}
-
-    ##################################################
-    #
-    #   Control
-    ###################################################
-
-    def control_discrete_actions(self, action):
-        """
-        working with LaneDetector
-        """
-        t = self.car.car.get_transform()
-        v = self.car.car.get_velocity()  # returns in m/sec
-        c = self.car.car.get_control()
-        w = self.car.car.get_angular_velocity()  # returns in deg/sec
-        a = self.car.car.get_acceleration()
-        self.params["speed"] = int(3.6 * math.sqrt(v.x**2 + v.y**2 + v.z**2))
-        self.params["steering_angle"] = w
-        # self.params["Steering_angle"] = steering_angle
-        self.params["Steer"] = c.steer
-        self.params["location"] = (t.location.x, t.location.y)
-        self.params["Throttle"] = c.throttle
-        self.params["Brake"] = c.brake
-        self.params["height"] = t.location.z
-        self.params["Acceleration"] = math.sqrt(a.x**2 + a.y**2 + a.z**2)
-
-        ## Applied throttle, brake and steer
-        curr_speed = int(3.6 * math.sqrt(v.x**2 + v.y**2 + v.z**2))
-        throttle = self.actions[action][0]
-        steer = self.actions[action][1]
-        # print(f"in STEP() {throttle = } and {steer = }")
-
-        target_speed = 30
-        brake = 0
-
-        if curr_speed > target_speed:
-            throttle = 0.45
-
-        self.car.car.apply_control(
-            carla.VehicleControl(throttle=throttle, steer=steer, brake=brake)
-        )
-
-    #####################################################################################
-    # ---   methods
-    #####################################################################################
-
-    def preprocess_image(self, img):
-        """
-        image is trimming from top to middle
-        """
-        ## first, we cut the upper image
-        height = img.shape[0]
-        image_middle_line = (height) // 2
-        img_sliced = img[image_middle_line:]
-        ## calculating new image measurements
-        # height = img_sliced.shape[0]
-        ## -- convert to GRAY
-        gray_mask = cv2.cvtColor(img_sliced, cv2.COLOR_BGR2GRAY)
-        ## --  apply mask to convert in Black and White
-        theshold = 50
-        # _, white_mask = cv2.threshold(gray_mask, 10, 255, cv2.THRESH_BINARY)
-        _, white_mask = cv2.threshold(gray_mask, theshold, 255, cv2.THRESH_BINARY)
-
-        return white_mask
-
-    def calculate_right_line(self, mask, x_row):
-        """
-        calculates distance from center to right line
-        This distance will be using as a error from center lane
-        """
-        ## get total lines in every line point
-        lines = [mask[x_row[i], :] for i, _ in enumerate(x_row)]
-
-        ### ----------------- from right to left
-        lines_inversed = [list(reversed(lines[x])) for x, _ in enumerate(lines)]
-        # print(f"{lines_inversed = }")
-        inv_index_right = [
-            np.argmax(lines_inversed[x]) for x, _ in enumerate(lines_inversed)
-        ]
-
-        index_right = [
-            mask.shape[1] - inv_index_right[x] if inv_index_right[x] != 0 else 0
-            for x, _ in enumerate(inv_index_right)
-        ]
-
-        return index_right
-
-    def preprocess_image_lane_detector(self, image):
-        """
-        image from lane detector
-        """
-        ## first, we cut the upper image
-        height = image.shape[0]
-        image_middle_line = (height) // 2
-        img_sliced = image[image_middle_line:]
-
-        kernel = np.ones((5, 5), np.uint8)
-        img_erosion = cv2.erode(img_sliced, kernel, iterations=1)
-        hsv = cv2.cvtColor(img_erosion, cv2.COLOR_RGB2HSV)
-        gray_hsv = cv2.cvtColor(hsv, cv2.COLOR_BGR2GRAY)
-        _, gray_mask = cv2.threshold(gray_hsv, 200, 255, cv2.THRESH_BINARY)
-
-        return gray_mask
-
-    def calculate_lane_centers_with_lane_detector(self, mask):
-        """
-        using Lane Detector model for calculating the center
-        """
-        lines = [mask[self.x_row[i], :] for i, _ in enumerate(self.x_row)]
-        index_left = [np.argmax(lines[x]) for x, _ in enumerate(lines)]
-        left_offset, right_offset = 20, 10
-        index_left_plus_offset = [
-            index_left[x] + left_offset if index_left[x] != 0 else 0
-            for x, _ in enumerate(index_left)
-        ]
-        second_lines_from_left_right = [
-            lines[x][index_left_plus_offset[x] :] for x, _ in enumerate(lines)
-        ]
-        ## ---------------- calculating index in second line
-        try:
-            index_right_no_offset = [
-                np.argmax(second_lines_from_left_right[x])
-                for x, _ in enumerate(second_lines_from_left_right)
-            ]
-        except:
-            secod_lines_from_left_right = [value for _, value in enumerate(self.x_row)]
-            index_right_no_offset = [
-                np.argmax(second_lines_from_left_right[x])
-                for x, _ in enumerate(second_lines_from_left_right)
-            ]
-
-        index_right_plus_offsets = [
-            index_right_no_offset[x] + left_offset + right_offset
-            for x, _ in enumerate(index_right_no_offset)
-        ]
-
-        index_right = [
-            right + left for right, left in zip(index_right_plus_offsets, index_left)
-        ]
-
-        centers = [
-            ((right - left) // 2) + left for right, left in zip(index_right, index_left)
-        ]
-
-        return centers, index_left, index_right
-
-
-# ==============================================================================
-# -- class RGB CameraManager -------------------------------------------------------------
-# ==============================================================================
-class CameraRGBSensor(object):
-    def __init__(self, parent_actor):
-        self.sensor = None
-        self._parent = parent_actor
-        self.front_rgb_camera = None
-
-        self.world = self._parent.get_world()
-        bp = self.world.get_blueprint_library().find("sensor.camera.rgb")
-        bp.set_attribute("image_size_x", f"640")
-        bp.set_attribute("image_size_y", f"480")
-        bp.set_attribute("fov", f"110")
-
-        self.sensor = self.world.spawn_actor(
-            bp,
-            carla.Transform(carla.Location(x=2, z=1.5), carla.Rotation(yaw=+00)),
-            attach_to=self._parent,
-        )
-        # We need to pass the lambda a weak reference to self to avoid circular
-        # reference.
-        weak_self = weakref.ref(self)
-        self.sensor.listen(lambda image: CameraRGBSensor._rgb_image(weak_self, image))
-
-    @staticmethod
-    def _rgb_image(weak_self, image):
-        """weakref"""
-        self = weak_self()
-        if not self:
-            return
-        image = np.array(image.raw_data)
-        image = image.reshape((480, 640, 4))
-        # image = image.reshape((512, 1024, 4))
-        image = image[:, :, :3]
-        # self._data_dict["image"] = image3
-        self.front_rgb_camera = image
-
-
-# ==============================================================================
-# -- Red Mask CameraManager -------------------------------------------------------------
-# ==============================================================================
-class CameraRedMaskSemanticSensor(object):
-    def __init__(self, parent_actor):
-        self.sensor = None
-        self._parent = parent_actor
-        self.front_red_mask_camera = None
-
-        self.world = self._parent.get_world()
-        bp = self.world.get_blueprint_library().find(
-            "sensor.camera.semantic_segmentation"
-        )
-        bp.set_attribute("image_size_x", f"640")
-        bp.set_attribute("image_size_y", f"480")
-        bp.set_attribute("fov", f"110")
-
-        self.sensor = self.world.spawn_actor(
-            bp,
-            carla.Transform(carla.Location(x=2, z=1.5), carla.Rotation(yaw=+00)),
-            attach_to=self._parent,
-        )
-        # We need to pass the lambda a weak reference to self to avoid circular
-        # reference.
-        weak_self = weakref.ref(self)
-        self.sensor.listen(
-            lambda image: CameraRedMaskSemanticSensor._red_mask_semantic_image_callback(
-                weak_self, image
-            )
-        )
-
-    @staticmethod
-    def _red_mask_semantic_image_callback(weak_self, image):
-        """weakref"""
-        self = weak_self()
-        if not self:
-            return
-        image.convert(carla.ColorConverter.CityScapesPalette)
-        array = np.frombuffer(image.raw_data, dtype=np.dtype("uint8"))
-        array = np.reshape(array, (image.height, image.width, 4))
-        array = array[:, :, :3]
-        array = array[:, :, ::-1]
-
-        hsv_nemo = cv2.cvtColor(array, cv2.COLOR_RGB2HSV)
-
-        if (
-            self.world.get_map().name == "Carla/Maps/Town07"
-            or self.world.get_map().name == "Carla/Maps/Town04"
-            or self.world.get_map().name == "Carla/Maps/Town07_Opt"
-            or self.world.get_map().name == "Carla/Maps/Town04_Opt"
-        ):
-            light_sidewalk = (42, 200, 233)
-            dark_sidewalk = (44, 202, 235)
-        else:
-            light_sidewalk = (151, 217, 243)
-            dark_sidewalk = (153, 219, 245)
-
-        light_pavement = (149, 127, 127)
-        dark_pavement = (151, 129, 129)
-
-        mask_sidewalk = cv2.inRange(hsv_nemo, light_sidewalk, dark_sidewalk)
-        # result_sidewalk = cv2.bitwise_and(array, array, mask=mask_sidewalk)
-
-        mask_pavement = cv2.inRange(hsv_nemo, light_pavement, dark_pavement)
-        # result_pavement = cv2.bitwise_and(array, array, mask=mask_pavement)
-
-        # Adjust according to your adjacency requirement.
-        kernel = np.ones((3, 3), dtype=np.uint8)
-
-        # Dilating masks to expand boundary.
-        color1_mask = cv2.dilate(mask_sidewalk, kernel, iterations=1)
-        color2_mask = cv2.dilate(mask_pavement, kernel, iterations=1)
-
-        # Required points now will have both color's mask val as 255.
-        common = cv2.bitwise_and(color1_mask, color2_mask)
-        SOME_THRESHOLD = 0
-
-        # Common is binary np.uint8 image, min = 0, max = 255.
-        # SOME_THRESHOLD can be anything within the above range. (not needed though)
-        # Extract/Use it in whatever way you want it.
-        intersection_points = np.where(common > SOME_THRESHOLD)
-
-        # Say you want these points in a list form, then you can do this.
-        pts_list = [[r, c] for r, c in zip(*intersection_points)]
-        # print(pts_list)
-
-        # for x, y in pts_list:
-        #    image_2[x][y] = (255, 0, 0)
-
-        # red_line_mask = np.zeros((400, 500, 3), dtype=np.uint8)
-        red_line_mask = np.zeros((image.height, image.width, 3), dtype=np.uint8)
-
-        for x, y in pts_list:
-            red_line_mask[x][y] = (255, 0, 0)
-
-        # t_end = self.timer.time()
-        # self.time_processing += t_end - t_start
-        # self.tics_processing += 1
-
-        red_line_mask = cv2.cvtColor(red_line_mask, cv2.COLOR_BGR2RGB)
-        self.front_red_mask_camera = red_line_mask
-
-        # AutoCarlaUtils.show_image(
-        #    "states",
-        #    self.front_red_mask_camera,
-        #    600,
-        #    400,
-        # )
-        # if self.front_red_mask_camera is not None:
-        #    time.sleep(0.01)
-        #    print(f"self.front_red_mask_camera leyendo")
-        #    print(f"in _red_mask_semantic_image_callback() {time.time()}")
-
-
-# ==============================================================================
-# -- class LaneDetector -------------------------------------------------------------
-# ==============================================================================
-class LaneDetector:
-    def __init__(self, model_path: str):
-        torch.cuda.empty_cache()
-        self.__model: torch.nn.Module = torch.load(model_path)
-        self.__model.eval()
-
-    def detect(self, img_array: np.array) -> tuple:
-        with torch.no_grad():
-            image_tensor = img_array.transpose(2, 0, 1).astype("float32") / 255
-            x_tensor = torch.from_numpy(image_tensor).to("cuda").unsqueeze(0)
-            back, left, right = (
-                torch.softmax(self.__model.forward(x_tensor), dim=1).cpu().numpy()[0]
-            )
-
-        res, left_mask, right_mask = self.lane_detection_overlay(img_array, left, right)
-
-        return res, left_mask, right_mask
-
-    def lane_detection_overlay(
-        self, image: np.ndarray, left_mask: np.ndarray, right_mask: np.ndarray
-    ) -> tuple:
-        """
-        @type left_mask: object
-        """
-        res = np.copy(image)
-
-        cv2.erode(left_mask, (7, 7), 4)
-        cv2.dilate(left_mask, (7, 7), 4)
-
-        cv2.erode(right_mask, (7, 7), 4)
-        cv2.dilate(right_mask, (7, 7), 4)
-
-        left_mask = self.image_polyfit(left_mask)
-        right_mask = self.image_polyfit(right_mask)
-
-        # We show only points with probability higher than 0.07
-        res[left_mask > 0.1, :] = [255, 0, 0]
-        res[right_mask > 0.1, :] = [0, 0, 255]
-
-        return res, left_mask, right_mask
-
-    def image_polyfit(self, image: np.ndarray) -> np.ndarray:
-        img = np.copy(image)
-        img[image > 0.1] = 255
-
-        indices = np.where(img == 255)
-
-        if len(indices[0]) == 0:
-            return img
-        grade = 1
-        coefficients = np.polyfit(indices[0], indices[1], grade)
-
-        x = np.linspace(0, img.shape[1], num=2500)
-        y = np.polyval(coefficients, x)
-        points = np.column_stack((x, y)).astype(int)
-
-        valid_points = []
-
-        for point in points:
-            # if (0 < point[1] < 1023) and (0 < point[0] < 509):
-            if (0 < point[1] < image.shape[1]) and (0 < point[0] < image.shape[0]):
-                valid_points.append(point)
-
-        valid_points = np.array(valid_points)
-        polyfitted = np.zeros_like(img)
-        polyfitted[tuple(valid_points.T)] = 255
-
-        return polyfitted
-
-
-# ==============================================================================
-# -- New Car -------------------------------------------------------------
-# ==============================================================================
-class NewCar(object):
-    def __init__(self, parent_actor, init_positions, init=None):
-        self.car = None
-        self._parent = parent_actor
-        self.world = self._parent
-        vehicle = self.world.get_blueprint_library().filter("vehicle.*")[0]
-
-        if init is None:
-            pose_init = np.random.randint(0, high=len(init_positions))
-        else:
-            pose_init = init
-
-        location = carla.Transform(
-            carla.Location(
-                x=init_positions[pose_init][0],
-                y=init_positions[pose_init][1],
-                z=init_positions[pose_init][2],
-            ),
-            carla.Rotation(
-                pitch=init_positions[pose_init][3],
-                yaw=init_positions[pose_init][4],
-                roll=init_positions[pose_init][5],
-            ),
-        )
-
-        self.car = self.world.spawn_actor(vehicle, location)
-        while self.car is None:
-            self.car = self.world.spawn_actor(vehicle, location)
 
 
 ##################################################################################################
@@ -1064,6 +294,11 @@ class TrainerFollowLaneDQNAutoCarlaTF:
                 self.world.unload_map_layer(carla.MapLayer.Particles)
                 self.world.unload_map_layer(carla.MapLayer.Props)
 
+            ## LaneDetector Camera ---------------
+            # self.sensor_camera_lanedetector = LaneDetector(
+            #    "models/fastai_torch_lane_detector_model.pth"
+            # )
+
             ####################################################################################
             # FOR
             #
@@ -1105,10 +340,10 @@ class TrainerFollowLaneDQNAutoCarlaTF:
                 self.actor_list.append(self.sensor_camera_red_mask.sensor)
 
                 ## LaneDetector Camera ---------------
-                self.sensor_camera_lanedetector = LaneDetector(
-                    "models/fastai_torch_lane_detector_model.pth"
-                )
-                self.actor_list.append(self.sensor_camera_lanedetector)
+                # self.sensor_camera_lanedetector = LaneDetector(
+                #    "models/fastai_torch_lane_detector_model.pth"
+                # )
+                # self.actor_list.append(self.sensor_camera_lanedetector)
 
                 # print(
                 #    f"in main() {self.sensor_camera_rgb.front_rgb_camera} in {time.time()}"
@@ -1125,7 +360,7 @@ class TrainerFollowLaneDQNAutoCarlaTF:
                     self.new_car,
                     self.sensor_camera_rgb,  # img to process
                     # self.sensor_camera_red_mask, # sensor to process image
-                    self.sensor_camera_lanedetector,  # sensor to process image
+                    # self.sensor_camera_lanedetector,  # sensor to process image
                     self.environment.environment,
                 )
                 # env = gym.make(self.env_params.env_name, **self.environment.environment)
@@ -1216,6 +451,7 @@ class TrainerFollowLaneDQNAutoCarlaTF:
                         # in_best_epoch_training_time=best_epoch_training_time,
                     )
 
+                    """
                     AutoCarlaUtils.show_image(
                         "RGB",
                         self.sensor_camera_rgb.front_rgb_camera,
@@ -1228,7 +464,9 @@ class TrainerFollowLaneDQNAutoCarlaTF:
                         500,
                         50,
                     )
+                    """
 
+                    """
                     AutoCarlaUtils.show_image_only_right_line(
                         "front RGB",
                         # self.front_rgb_camera[(self.front_rgb_camera.shape[0] // 2) :],
@@ -1246,6 +484,7 @@ class TrainerFollowLaneDQNAutoCarlaTF:
                         env.drawing_lines_states,
                         env.drawing_numbers_states,
                     )
+                    """
 
                     # best episode
                     if current_max_reward <= cumulated_reward and episode > 1:
